finish full crud routes for server

+++ app/api/server_routes.py
@@ -1,4 +1,4 @@
-from flask import Blueprint, jsonify, request
+from flask import Blueprint, jsonify, request, redirect
 from flask_login import login_required
 from app.models import Server, db
 from app.forms import ServerForm
@@ -34,25 +34,36 @@ def server_create():
         return new_server.to_dict()
     return form.data.error
 
-@server_routes.route('/<int: serverId>', methods=['PATCH'])
+@server_routes.route('/<int: serverId>/update', methods=['GET', 'POST'])
 def update_server(serverId):
     """
     Update a server
     """
-    server = Server.query.get_or_404(serverId)
-    name = request.json['name']
-    preview_img = request.json['preview_img']
-    
-    server.name = name
-    server.preview_img = preview_img
-    
-    db.session.commit()
-    return server.to_dict() # double check this
+    server = Server.query.get_or_404(serverId) # query or 404 if not found
+    if request.method == 'POST':
+        if server:
+            db.session.delete(server)
+            db.session.commit()
+            
+            name = request.json['name']
+            preview_img = request.json['preview_img']
+            server = Server(id = serverId, name = name, preview_img = preview_img)
+            
+            db.session.add(server)
+            db.session.commit()
+            return server.to_dict()
+        return "Server does not exist"
 
-@server_routes.route('/<int: serverId>', methods=['DELETE'])
+@server_routes.route('/<int: serverId>/delete', methods=['GET', 'POST'])
 def delete_server(serverId):
+    """ 
+    Delete a route
+    """
     server = Server.query.get_or_404(serverId)
-    db.session.delete(server)
-    db.session.commit()
-    
-    return #something?
+    if request.method == 'POST':
+        if server:
+            db.session.delete(server)
+            db.session.commit()
+            return redirect('/api/servers')
+    return "Server does not exist"
+
